// src/app/services/models/departement-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DepartementShortDto } from '../models/departement-short-dto';
import { EmployeeShortDto } from '../models/employee-short-dto';
export interface DepartementDto {
  active?: boolean;
  code?: string;
  description?: string;
  id?: number;
  manager?: EmployeeShortDto;
  name?: string;
  parentDepartement?: DepartementShortDto;
}